Move thresholds into report package
