+++ src/main/scala/de/johoop/jacoco4sbt/JacocoItPlugin.scala
@@ -66,7 +66,6 @@ object JacocoItPlugin extends BaseJacocoPlugin with Merging {
           (coveredSources in ItJacoco).value,
           (classesToCover in ItJacoco).value,
           (jacocoSourceSettings in ItJacoco).value,
-          (thresholds in ItJacoco).value,
           (streams in ItJacoco).value
         ),
         jacocoReport in ItJacoco := ((jacocoReport in ItJacoco) dependsOn conditionalMerge).value
