+++ src/sbt-test/jacoco4sbt/simple/build.sbt
@@ -1,5 +1,3 @@
-import de.johoop.jacoco4sbt.Thresholds
-
 name := "jacocoTest"
 organization := "com.navetas"
 
@@ -9,10 +7,14 @@ scalacOptions ++= Seq("-deprecation", "-optimize", "-unchecked", "-Xlint", "-lan
 libraryDependencies += "org.scalatest" %% "scalatest" % "3.0.4" % "test"
 
 fork in Jacoco := false
-thresholds in Jacoco := Thresholds(
-  instruction = 100,
-  method = 100,
-  branch = 100,
-  complexity = 100,
-  line = 100,
-  clazz = 100)
+
+jacocoReportSettings in Jacoco := JacocoReportSettings()
+  .withThresholds(
+    JacocoThresholds(
+      instruction = 100,
+      method = 100,
+      branch = 100,
+      complexity = 100,
+      line = 100,
+      clazz = 100)
+  )
