+++ src/test/scala-sbt-0.13/de/johoop/jacoco4sbt/TestCounters.scala
@@ -12,7 +12,7 @@
 
 package de.johoop.jacoco4sbt
 
-import de.johoop.jacoco4sbt.report.{JacocoReportSettings, JacocoSourceSettings}
+import de.johoop.jacoco4sbt.report.{JacocoReportSettings, JacocoSourceSettings, JacocoThresholds}
 import org.jacoco.core.analysis.{IBundleCoverage, ICounter}
 import org.mockito.Mockito.{mock, when}
 import sbt.Keys.TaskStreams
@@ -26,10 +26,11 @@ class TestCounters {
 
   private val report =
     new Report(
-      thresholds = Thresholds(instruction = 35, method = 40, branch = 30, complexity = 35, line = 50, clazz = 40),
       streams = mockStreams,
       sourceSettings = JacocoSourceSettings(),
-      reportSettings = JacocoReportSettings(),
+      reportSettings = JacocoReportSettings(
+        thresholds =
+          JacocoThresholds(instruction = 35, method = 40, branch = 30, complexity = 35, line = 50, clazz = 40)),
       reportDirectory = null,
       executionDataFiles = Nil,
       classDirectories = Nil,
@@ -51,9 +52,15 @@ class TestCounters {
   }
 
   def stubCounters(missedCount: Seq[Int], totalCount: Seq[Int], coveredRatio: Seq[Double]): Unit = {
-    missedCount.foldLeft(when(mockICounter.getMissedCount)) { (acc, c) => acc.thenReturn(c) }
-    totalCount.foldLeft(when(mockICounter.getTotalCount)) { (acc, c) => acc.thenReturn(c) }
-    coveredRatio.foldLeft(when(mockICounter.getCoveredRatio)) { (acc, c) => acc.thenReturn(c) }
+    missedCount.foldLeft(when(mockICounter.getMissedCount)) { (acc, c) =>
+      acc.thenReturn(c)
+    }
+    totalCount.foldLeft(when(mockICounter.getTotalCount)) { (acc, c) =>
+      acc.thenReturn(c)
+    }
+    coveredRatio.foldLeft(when(mockICounter.getCoveredRatio)) { (acc, c) =>
+      acc.thenReturn(c)
+    }
   }
 
   def checkCounter(required: Double): Boolean = {
